Update patch set 3
Patch Set 3:

Mike, I think your last commit was an empty commit, no changes, just the timestamp of the commit message.

I wounder if you miss to submit the 2 small little changes that you had pending.   If I recall correctly, these are the only 2 things to take care of to close this review:

1. Per Chenfei, remove the 400 line in swagger.
2. Per Ram, add the 403 status line.

Once you add that, I think this should be ready to be merged.

Patch-set: 3
Label: Code-Review=0