Added HTTPS and CADI/AAF Support for PDP-X
This is a draft intended for feedback.  I have been stuck on the https
imlementation. Junit test are not passing due to a SSL Handshake error
with message "no cipher suites in common".  When stepping through the
code it seems to me that everything is get configured properly for HTTPS
when creating the Invocation.Builder object, however it fails when
executing invocationBuilder.get(HealthCheckReport.class) on line 129 of
TestXacmlPdpRestServer.java and likewise
for StatisticsReport.

Change-Id: Ie1a6225459b3ce235cd73828ccddec04c690f5fd
Issue-ID: POLICY-1436
Signed-off-by: Michael Mokry <michael.mokry@att.com>