Update patch set 3
Patch Set 3:

Hi Josh, just an FYI to keep you in the loop, given the tight deadlines we have ahead of us, we are being more lenient than usual with all the reviews that relate to the new architecture components for team members.   Having said that, the contributor should consider all comments, but is also given the flexibility of attending them in the same or in subsequent reviews.    This also means that there is the possibility that sonars and bugs that could be introduced, but the main benefit is that development would progress swiftfly.  You're doing a great job with the reviews, and your comments are much more than appreciated, but I wanted to keep you informed of this strategy we are implementing for the near term, since there may be some conversations that you many not be aware of from the scrums.

Patch-set: 3