Update patch set 1
Patch Set 1:

I'm not sure what you mean. Test-driven development is a standard agile practice, and given that I'm familiarizing myself with the codebase as I attempt to contribute, it seemed the best way to make quick progress. I know you are under substantial pressure, so figured it would be worth trying to lend a hand on the guard stuff. I've got time starting later this morning in which I can flesh this test out and write the associated application code (e.g., in StdMetadataTranslator.java). Please let me know if you'd like me to proceed, or if you'd prefer to write yourself and have me extend that code this upcoming week. Also, if you'd like to discuss further, feel free to try my cell 516 816 2470 (if no response, just leave a message or text and I'll call you back).

Patch-set: 1