Adding optimization application
Created Optimization application and created a translator
for it. The translator makes an assumption that OOF wants
to query on policyScope and policyType properties.

Rearranged some of the test code for re-usability.

There is still some inconsistencies and duplicate code
that I can take care of.

Added guard policy creation. But right now it isn't
completely working yet.

Need to release xacml 2.0.1 to fix an annotation
glitch for null values.

Issue-ID: POLICY-1273
Change-Id: Ia3dbfa992bbe0cbb5aa294c38aa2aff430a3230a
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>