Update patch set 1
Patch Set 1: Code-Review-2

This test is not needed here. The parsing of these files is done in policy/models repo.
You are testing whether a file is correct. the api will take care of that for us, there are tools in place to do so.
These junit will test the functionality of the guard application.

Patch-set: 1
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2