Fix comments and add debugging
Added some fixes from previous review and some debugging
for the JUnits. The merge failed in the JUnits and I
could not see why. I think the wrong request was sent, but
I'm not sure why the JUnits woulnd't fail locally or in
the verify job. We need to do a lot of testing with these
policy translations.

Also fixed the guard policies to use correct VF Module Create
recipe.

Issue-ID: POLICY-1273
Change-Id: I813bcca92436d356c43da5613eab7ba7746a4e3d
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>