Update patch set 8
Patch Set 8:

(3 comments)

min/max is failing b/c tests are currently independent. specifically, in test3 a frequency guard is loaded. When test4 runs it adds a second min/max guard, but the frequency guard evaluates to deny and the root combining algorithm used is permit-unless-deny, so deny results and test4 fails.

Patch-set: 8