Update patch set 8
Patch Set 8:

Ah, very welcome, got it. In that case, the important point is that test4 is currently failing on the frequency guard, not the minmax guard (I modified the test locally and confirmed that test4 passed when only the minmax guard is loaded). That info should help you restructure, though it strikes me as a bit tricky since the frequency guard is going to deny anytime you've got 1 or more successful actions and the minmax is going to deny anytime you've got less than 1 successful action. Thinking out loud, I guess one option is changing the frequency guard to have a limit of 2 instead of 1, calling insertOperationEvent twice, running test3, clearing the db (but not the PDP) after test 3, then and calling insertOperationEvent once in test4. Anyway, I'll stay posted for additional code submssions, etc.

Patch-set: 8