Update patch set 3
Patch Set 3:

> I am just saving my work here as I'm about to run to the airport.
 > Mike - please pay attention to the application manager static class
 > and let me know about any comments. Will fix the JUnit test a bit
 > later today and work on all this on the plane. Suggestions welcome.

Looks good to me.. I am wondering, since I used your locateExistingPolicyTypes() method in the XacmlPdpMessage class I created to structure messages to the PAP, if we can just add a piece of code here that would add the supportedPolicyTypes to the policyTypeIdents list used in registration and PdpStatus messages to the PAP.  Then I can remove that method and just get the policyTypeIdents from here rather than have redundant code in the XacmlPdpMessage class.  If you agree, I can make the change and submit it with my next patch to my review for registration today.

Patch-set: 3