Test decision from main entry
Tests a decision upon startup. This also updates the use of
ToscaPolicyTypeIdentifier for all the applications.

Very basic packaging for applications and their properties.

Added cleaning of unused imports to remove sonar issues.

Added production persistence.xml file to guard application.
Not sure if we need a copy in the application/common resource
directory.

Issue-ID: POLICY-1440
Change-Id: If96eef5a9e0a7c6cc5461c0bdb6f0cd708cc41bb
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>