Update patch set 3
Patch Set 3:

> Have you had a chance to look at drools-applications guard code
 > yet? I changed the attribute id's etc and db name. It will need to
 > be updated along with the drools rules

I haven't had a chance to start on the drools-application guard code, but can take care of it later today or tomorrow depending on how things go. To confirm: the plan is to update the drools-application codebase to match the xacml-pdp codebase (e.g., so drools-application uses the same identifiers, pips, tests, etc.), correct?

Patch-set: 3