Update patch set 9
Patch Set 9:

Yes Josh - same conclusion its the date in range that is probably failing. Possibly a bug with date/time environment attributes from the xacml engine on the jenkins VM's. I may have to put more debugging information in the xacml engine to see what the problem is.

I saw the xacml properties problem because it is dumping properties out, and all of a sudden all the properties were missing except for the new policy. That is not good, as we want to use our policy finder and pip settings.

Thank you for helping to take a look at it last night, I appreciate your effort.

I will continue this morning - but may comment out some of the tests temporarily in order to get your code moving. When we find a local environment showing the same problems, then let's make an effort to debug it if possible.

Patch-set: 9