Fix the xacml-pdp CSIT failures
Looks like CSIT uses docker to install xacml-pdp so a change in OOM is
not necessary to fix this problem.  I took a look at the PAP script and
found some small differenences and made those changes in xacml-pdp.  I
have not been able to successfully test this as I spent all night trying
to figure out how to execute it in my VM with and without arguments. I
added some debugger code in as well to get more information in the logs
and see if the property file is actually being found.

Change-Id: Ib6ca95edab4483aa4e65efca73a9f58342cc1ec0
Issue-ID: POLICY-1449
Signed-off-by: Michael Mokry <michael.mokry@att.com>