Update patch set 9
Patch Set 9:

Let's see how we can debug this. For the moment, I cannot change the xacml github code this close to code freeze. But perhaps Josh, you could take a look at the xacml github code and how it generates the environment attribute values. Copy that code into a JUnit in your guard test code so we can see what that jenkins environment is calculating.

I will on a weekend soon go into the xacml github and see about adding more TraceEngine details. Obviously the Condition is returning false and most likely it is the timerange check. But it would be good to know the values. I don't know when I can get to that. But if you do the above to start, that would help. Maybe there is a bug to fix there. I'm sure it is some sort of a UTC issue.

Patch-set: 9