Update patch set 7
Patch Set 7: Code-Review+1

At this point - enough is there. Mike, we still need to cleanup the main application and determine why things are left hanging around. The statics are not implemented correctly.
I need to move on to the Blacklist guard. So this review needs to get in.

Patch-set: 7
Label: Code-Review=+1