Update patch set 3
Patch Set 3:

I guess I am not sure about the destination location where to place it, you guys know better, once the container is running.  The master copy as I understand it from Mike is placed outside of the container and from outside ../etc/mounted dir (same as the config.json) is the name of the directory that is made available to the container on the helm charts.

Patch-set: 3