Update patch set 3
Patch Set 3:

> I guess I am not sure about the destination location where to place
 > it, you guys know better, once the container is running.  The
 > master copy as I understand it from Mike is placed outside of the
 > container and from outside ../etc/mounted dir (same as the
 > config.json) is the name of the directory that is made available to
 > the container on the helm charts.

When we had the discussion in Rocketchat I posted the tree layout that is contained in the tarball:
policy/xacml-pdp/packages/policy-xacmlpdp-tarball/src/main/resources/apps.
The config JSON file has an entry that points to the apps directory.
Mike knows this because he put the package together. So we have to determine how to replicate the apps directory for Helm charts and then modify ONLY the xacml.properties under the apps/guard directory.

Patch-set: 3