Update patch set 3
Patch Set 3:

This should work fine I believe Mike.   For the sake of time, I suggest adopting this simplified version so to have it available for release tomorrow.   The code will only be executed in an OOM installation:
-----
if [[ -f ${POLICY_HOME}/etc/mounted/xacml.properties ]]; then
    cp -f “${POLICY_HOME}"/etc/mounted/xacml.properties  "${POLICY_HOME}"/apps/guard/
fi
-----

A couple notes you can use POLICY_HOME or POLICY_PDPX_HOME they should point to the same.   The base Dockerfile uses POLICY_HOME so it is available and should work.

“${POLICY_HOME}/etc/mounted/xacml.properties” is fixed and will only appear in an oom installation, so the copy will only apply in those cases and that's what you want.

Looking at the layout structure of pdp-x that Pam referred to and in the container, the guard location is ${POLICY_HOME}”/apps/guard/ and you're providing this external xacml.properties.

I think that's the path of least resistence so we can release tomorrow.   After that you could add any enhancements you may want to it.

Patch-set: 3