Update patch set 6
Patch Set 6:

Ok I think we are still waiting for Mike to confirm he is done as well as Jorge to finish review. Once done, I would like this cherry-picked to Dublin. Just in case we need to provide a fix for this sometime before El Alto release. Thanks!

Patch-set: 6