Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1

I agree, but can we do that work in a different Jira since its not actually a bug but an enhancement to the feature.  This fixes the the current bug with creating the table, can we merge this and expand it to be configurable in a separate review with a new Jira ticket?  If not no biggie I can work on it here too but thought it may be better tracked as two separate things.

Patch-set: 1