Update patch set 1
Patch Set 1: -Code-Review

> > Patch Set 1: Code-Review-1
 > 
 > I agree, but can we do that work in a different Jira since its not
 > actually a bug but an enhancement to the feature.  This fixes the
 > the current bug with creating the table, can we merge this and
 > expand it to be configurable in a separate review with a new Jira
 > ticket?  If not no biggie I can work on it here too but thought it
 > may be better tracked as two separate things.

I am open to any solution as long as the work is done at some point soon.
I think though this work is part of the bug and not an enhancement. In addition, at some point or another, the security team will flag us for hardcoding these values (username/passwords/hosts) in the image. 

The point I am trying to make is that in CSIT tests, this will cause an error (the table will not be created), so to me that's the same but.   So .. before this fix it was working for CSIT (table created) but not for OOM deployments.   Now with this fix, it will work for OOM deployments, but not for CSITs (no table created).  In essence, the fault was flipped from OOM to CSIT side.

The problem perhaps is asymptomatic in CSIT tests, as there may be no tests that will execute any tests that requires this table around.

Patch-set: 1
Label: Code-Review=0