Update patch set 1
Patch Set 1:

> (1 comment)

This code doesn't seem to be working like that, only checks for "abbrev" and discards what was passed in.   

In any case, I would assume query parameters should be fairly limited in its usage, and should be specified in the XacmlPdpRestController in the swagger apidoc to show up it generated UIs.

Patch-set: 1