Update patch set 1
Patch Set 1: -Code-Review

I will let you decide which way you want to go.  In my opinion, the point of jaxrs instead of using directly the low level HttpServletRequest is to support query parameters with annotations which also serve as self-documentation.  

If we are to support the query parameters that can be validated a priori in the code, would make sense.  But the code as currently is does not work as such as I mentioned before.

Patch-set: 1
Label: Code-Review=0