More examples of optimization policies and cleanup
Fixed a couple of sonar issues - log exception and do not
nest more 3 if-else-try.

Cleaned up the JUnit test to make debugging a bit easier.

Added more examples for testing optimization tests.

Moved the target back into the Policy and kept the Condition
on the Rule. Works exactly the same, just a bit cleaner and
one less rule to deal with.

Issue-ID: POLICY-2066
Change-Id: Ife28dc2ce959dcf1fb8ca72061ebc5dca862a7f4
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>