Implement closest match algorithm
Needed to add more obligations, which make it easier to
scan through obligations and be able to pull information
about the policy.

Adding weight as an obligation for Optimization policies in
order to implement "closest match" algorithm.

Moved Obligation to a support class.

Added JUnit code coverage on the translator classes.

Split some methods up to reduce complexity via sonar.

Issue-ID: POLICY-2066
Change-Id: Ibb13d2dc0a63ab2a6d585b0697a0c1d129fa8f7b
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>