Update patch set 8
Patch Set 8: Code-Review+1

I think I will defer some of the work to Friday - I need to get this in. I can re-factor the OnapObligation into some sub-classes for each translator then and ensure that code coverage for each JUnit is 100%. Most are right now, but a couple need a bit of work.

Patch-set: 8
Label: Code-Review=+1