Update patch set 3
Patch Set 3:

(2 comments)

Does this change mean that an XACML PDP must always be in a PDP group that's in the configuration? Wouldn't it be better to have the PDP group in the configuration as the default setting and allow the PAP to set the PDP group if the PDP group is not set in the config?

Also I can't see where the PDP group is set by the PAP in the current code.

Patch-set: 3