Update patch set 2
Patch Set 2:

> > Patch Set 2:
 > >
 > > Should the status class then remove the supportedPolicyTypes
 > field? Or does that need to stay?
 > 
 > I don't see PdpStatus supportedPolicyTypes being used anywhere else
 > in xacml-pdp. It's also only used in the PdpStatus constructor in
 > policy/models.. I'm not sure if other projects depend on this field
 > though?

SupportedPolicyTypes will be removed from PdpStatus class (in policy/models) once xacml & drools stop referring to that. 
But the PDP is still expected to hold the supportedPolicyTypes that it supports. Its just that PDPs don't need to report it in the PdpStatus message.

The next story "resync of PAP & PDP when supportedPolicyTypes of PdpSubgroup is changed" will need the PDPs to verify their list of supportedPolicyTypes with the updated one provided by PAP in order to decide that they still want to be part of the updated group or want to leave it.

Patch-set: 2