Update patch set 2
Patch Set 2:

It is a new implementation option in which endpoint accepts an application/json or application/xacml+xml payload. Typically it is a string passed into the provider. The native application and its translator will translate the string into xacml request based on the content-type.

I use JsonRequestTranslator.load(string) and DOMRequest.load(string) inspired by their usage in XACMLPdpServlet (github xacml). @Pam, please let me know if it is what you are looking for. If so, I'll add in the junits.

Patch-set: 2