Update patch set 7
Patch Set 7:

The native application junit that made me most nervous succeeds now with changing the encoding scheme to Base64. As xacml policy xml always starts with "<?xml" which will be URL encoded to "%...". "%" can't be the first character so I added a whitespace to it yesterday. The whitespace caused the error. 

More unit tests are coming.

Patch-set: 7
Label: Code-Review=0