Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> (1 comment)
> 
> > This is the consolidated version of the review which I abandoned
>  > https://gerrit.onap.org/r/#/c/policy/xacml-pdp/+/101557/
>  > 
>  > All the previous comments have been addressed and integrated in
>  > this review. It can build now.
> 
> Congratulations on getting it to build successfully!
> 
> Question: Should responses be encoded using the same media type as the request?  Or should they simply be test/plain or simple application/json?

I'm going to force that if they send XML, we return XML. The Content-Type and Accepts have to match. No mixing, although it should be possible. We can tackle that in another release. But I can't see someone want to send XML and get JSON back. If they are using one or the other most likely they would not mix.

Patch-set: 1