use shutdown to clear handle leak
The XACML github was released with a new method to allow
context factories and PIP engines to release any handles
before releasing the PDP engine. This review includes
that artifact and adds tests to support it.

In addition, added more tests to get code coverage over
90% for both PIPs in ONAP.

Some cleanup in the Matchable types based on last review.

Issue-ID: POLICY-2242
Change-Id: I312f06380ff4d2e16bcfd25b6d1f36ce5dd030e6
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>