Fix xacml-pdp csit test cases
In the recent runs, it was found that get statitcs might take few seconds longer after the actual deployment of policies. So, increasing the "max" wait time for the test cases to 1 min. And it won't slow down the tests as the check will be made every 15 sec to verify if the assertions are passing. The moment they pass, it will move on.

Issue-ID: POLICY-3257
Change-Id: I505c6e95e78a35b8bf1ee9921efbfcb4bd4874fb
Signed-off-by: Ram Krishna Verma <ram_krishna.verma@bell.ca>