Use protected methods vs private
Application and common code in xacml-pdp is meant to be
available for developers to re-use and/or extend as desired.

Private methods are unnecessary at this time, but could be
re-visited in a later release.

For Honolulu, simply moving to protected to keep functionality
the same. Possibly for Istanbul we can make more changes to
support mixing of ONAP application with custom applications
during runtime.

Issue-ID: POLICY-3305
Change-Id: I6db23bda3c81fdbbf86e39cc2cd2e706d8bee63d
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>
(cherry picked from commit 8e0aa0ffa5cef537de1c841acf38300832c52a56)
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>