Leave xacml-pdp REST server always running
Liveness checks depend on the xacml-pdp REST server being available.
However, the REST server is only running when PAP puts it in an active
state.  Modified the code to always leave it running.

In a subsequent review, we should split the healthcheck REST service
onto its own port so it can be left running, while the other services
are started/stopped in response to PAP active/passive requests.

Issue-ID: POLICY-3531
Change-Id: I412064abaf91bb966d40adc46cee771b3a0a5dfc
Signed-off-by: Jim Hahn <jrh3@att.com>