Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> It would be nice to have prometheus counts for the decisions also. That is very useful for an operations team to see if unusual amounts of denies are happening.

Yes, we talked about the support of metrics on a per application basis that should capture them.   Or may be we can just start a little simpler and create counters for the stats that you already have on prometheus, a straight mapping.   Isaac, can we arrange a meeting with Pam so we can collect more in detail feedback for next steps?

Patch-set: 3
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>